+++ src/app/base-page/base-page.component.ts
@@ -4,9 +4,10 @@ import {
   ElementRef,
   EventEmitter,
   Input,
-  OnDestroy,
-  Output, TemplateRef,
-  ViewChild, OnInit
+  OnInit,
+  Output,
+  TemplateRef,
+  ViewChild
 } from '@angular/core';
 import {FormService} from '../services/form.service';
 import fhir from 'fhir/r4';
@@ -19,9 +20,9 @@ import {FetchService} from '../services/fetch.service';
 import {FhirService} from '../services/fhir.service';
 import {FhirServersDlgComponent} from '../lib/widgets/fhir-servers-dlg/fhir-servers-dlg.component';
 import {FhirSearchDlgComponent} from '../lib/widgets/fhir-search-dlg/fhir-search-dlg.component';
-import { PreviewDlgComponent } from '../lib/widgets/preview-dlg/preview-dlg.component';
+import {PreviewDlgComponent} from '../lib/widgets/preview-dlg/preview-dlg.component';
 import {AppJsonPipe} from '../lib/pipes/app-json.pipe';
-import {Util} from '../lib/util';
+import {GuidingStep, Util} from '../lib/util';
 import {MatDialog} from '@angular/material/dialog';
 import {FhirExportDlgComponent} from '../lib/widgets/fhir-export-dlg/fhir-export-dlg.component';
 import {LoincNoticeComponent} from '../lib/widgets/loinc-notice/loinc-notice.component';
@@ -39,7 +40,7 @@ export class BasePageComponent implements OnInit {
 
   private unsubscribe = new Subject<void>();
   @Input()
-  guidingStep = 'home'; // 'choose-start', 'home', 'item-editor'
+  guidingStep: GuidingStep = 'home'; // 'choose-start', 'home', 'item-editor'
   startOption = 'scratch';
   importOption = '';
   questionnaire: fhir.Questionnaire = null;
@@ -102,7 +103,7 @@ export class BasePageComponent implements OnInit {
       console.log('Saved');
     });
 
-    formService.guidingStep$.subscribe((step) => {this.guidingStep = step;});
+    formService.guidingStep$.subscribe((step: GuidingStep) => {this.guidingStep = step;});
     FormService.lformsLoaded$.subscribe({error: (error) => {
       this.lformsErrorMessage = `Encountered an error which causes the application not to work properly. Root cause is: ${error.message}`;
     }});
@@ -178,7 +179,13 @@ export class BasePageComponent implements OnInit {
   addWindowListeners() {
     if (this.openerUrl) {
       const msgListener = (event) => {
-        const message = event.data;
+        let message = null;
+        try {
+          message = JSON.parse(JSON.stringify(event.data)); // Sanitize input.
+        } catch (e) {
+          console.error(`Unrecognized input from ${event.origin}`);
+          return;
+        }
         const parentUrl = this.formService.windowOpenerUrl;
         if(!parentUrl.startsWith(event.origin)) {
           return;
@@ -196,7 +203,7 @@ export class BasePageComponent implements OnInit {
             break;
 
           default:
-            console.log(`Received a message from ${parentUrl}: type = ${event.data?.type}`);
+            console.log(`Received a message from ${parentUrl}: type = ${message?.type}`);
             break;
         }
       }
@@ -276,7 +283,7 @@ export class BasePageComponent implements OnInit {
    * Switch guiding step
    * @param step - a
    */
-  setStep(step: string) {
+  setStep(step: GuidingStep) {
     this.formService.setGuidingStep(step);
     this.formService.autoSave('state', step);
   }
@@ -293,6 +300,7 @@ export class BasePageComponent implements OnInit {
   onContinue() {
     if(this.startOption === 'from_autosave') {
       let state = this.formService.autoLoad('state');
+      state = Util.isGuidingStep(state) ? state : 'fl-editor';
       state = state === 'home' ? 'fl-editor' : state;
       this.formService.setGuidingStep(state);
       this.setQuestionnaire(this.formService.autoLoadForm());
@@ -347,9 +355,16 @@ export class BasePageComponent implements OnInit {
    */
   onFileSelected(event) {
     const loadFromFile = () => {
-      const fileReader = new FileReader();
-      const selectedFile = event.target.files[0];
+      const file = event.target.files[0];
+      const selectedFile = Util.validateFile(file);
+
+      if(!selectedFile) {
+        this.showError(`Invalid file name: ${file?.name}`);
+        return;
+      }
+
       event.target.value = null; //
+      const fileReader = new FileReader();
       fileReader.onload = () => {
         setTimeout(() => {
           this.setStep('fl-editor');
@@ -364,6 +379,7 @@ export class BasePageComponent implements OnInit {
       fileReader.onerror = () => {
         this.showError(`Error occurred reading file: ${selectedFile.name}`);
       }
+
       fileReader.readAsText(selectedFile, 'UTF-8');
     };
 
@@ -497,6 +513,7 @@ export class BasePageComponent implements OnInit {
    */
   close() {
     if(this.openerUrl) {
+      // window.opener.postMessage({type: 'closed', questionnaire: Util.convertToQuestionnaireJSON(this.formValue)}, this.openerUrl);
       window.close();
     }
     else {