+++ src/app/services/form.service.ts
@@ -26,10 +26,11 @@ import ngxFlSchema from '../../assets/ngx-fl.schema.json5';
 import flLayout from '../../assets/fl-fields-layout.json5';
 // @ts-ignore
 import itemEditorSchema from '../../assets/item-editor.schema.json5';
-import {Util} from '../lib/util';
+import {GuidingStep, Util} from '../lib/util';
 import {FetchService} from './fetch.service';
 import {TerminologyServerComponent} from '../lib/widgets/terminology-server/terminology-server.component';
 import {ExtensionsService} from './extensions.service';
+
 declare var LForms: any;
 
 @Injectable({
@@ -39,7 +40,7 @@ export class FormService {
   static _lformsLoaded$ = new Subject<string>();
 
   private _loading = false;
-  _guidingStep$: Subject<string> = new Subject<string>();
+  _guidingStep$: Subject<GuidingStep> = new Subject<GuidingStep>();
   _formReset$: Subject<void> = new Subject<void>();
   _formChanged$: Subject<SimpleChange> = new Subject<SimpleChange>();
   _advPanelState = {
@@ -60,6 +61,7 @@ export class FormService {
 
   fetchService = inject(FetchService);
   formLevelExtensionService = inject(ExtensionsService);
+
   constructor(private modalService: NgbModal, private http: HttpClient) {
     [{schema: ngxItemSchema as any, layout: itemLayout}, {schema: ngxFlSchema as any, layout: flLayout}].forEach((obj) => {
       if(!obj.schema.definitions) {
@@ -249,7 +251,7 @@ export class FormService {
    * Inform the listeners of change in step.
    * @param step
    */
-  setGuidingStep(step: string) {
+  setGuidingStep(step: GuidingStep) {
     this._guidingStep$.next(step);
   }